+++ docs/examples/extension/Subscriber/UserSubscriber.php
@@ -11,7 +11,7 @@
  * that were distributed with this source code.
  */
 
-namespace Example\Event\Subscriber;
+namespace Example\Subscriber;
 
 use Eventum\Event\SystemEvents;
 use Symfony\Component\EventDispatcher\EventSubscriberInterface;